+++ src/lufah/commands/core/create_group.py
@@ -8,3 +8,10 @@ async def do_create_group(args: argparse.Namespace):
     client = args.client
     await client.connect()
     await client.create_group(client.group)
+
+
+async def do_delete_group(args: argparse.Namespace):
+    'Delete group if it exists, is not "", is paused, and has no units.'
+    client = args.client
+    await client.connect()
+    await client.delete_group(client.group)